Find the majority master node in an IoT setup (Course: Design and Analysis of Algorithms)
In this project, an IoT firm has designed a master-slave architecture where multiple sensors are responding to a single master node. These devices have different ID and storage capacity, but the data collected by them is going to a specific master node. The firm has not planned for load balancing. As the number of sensors/devices increased and the firm engineers kept on assigning the devices to a master nodes based on the maximum capacity, it created a unique problem where some of the  master nodes are overloaded and some of them are under loaded.

Skills used - Lists, Divide and conquer algorithm, Recursion