Find the routes between specified nodes in graph data structure ( Course: Design and Analysis of Algorithms )
In this project, a rectangular city area is surrounded by 4 different cities on each of its border. The rectangular city area is comprised of a grid system where altitude of different nodes are stored using python program with the help of graph data structures. The goal is to use graph data structure and perform BFS or DFS to find the route between 2 nodes and the route between 2 cities. Here we are using 2D matrix list to store the data in graph.

Skills used - Graphs, Operations on adjacency list graph, 2D matrix, List operations, Searching techniques, Data structures