Store the power grid data in graph data structure ( Course: Data Structures and Algorithms )
In this project, a power grid system where data of connected houses with different propagators are stored using python program with the help of graph data structures. The goal is to use graph data structure and perform certain operations such addition, removal of houses and propagators from the designed structure. Here we are using dictionary as an adjacency list to store the data in graph.

Skills used - Graphs, Operations on adjacency list graph, house creation, propagation creation, house removal